fixed some invalid types of arguments

+++ app/Orm/Castle/Castle.php
@@ -78,7 +78,7 @@ class Castle extends \Nextras\Orm\Entity\Entity {
   protected function getterRepairPrice(): int {
     if($this->hp >= 100) return 0;
     if($this->level === 1) $multiplier = 1; else $multiplier = ($this->level - 1) * 10 / 100 + 1;
-    $basePrice = self::BASE_REPAIR_PRICE * $multiplier * (100 - $this->hp);
+    $basePrice = (int) (self::BASE_REPAIR_PRICE * $multiplier * (100 - $this->hp));
     return (int) ($basePrice - $this->eventsModel->calculateRepairingDiscount($basePrice));
   }
